Small refactor and removed pointers
There is nothing being modified in each of the funcs defined on
CustomerIO and CustomerIOError, so I've removed the pointers. Also moved one
or two things around with to keep consistant with the rest of your code.

There are things here that I'd like to have changed, but have not, due to them
having an effect on the public API of this package. CustomerIOError being in a package
named customerio will cause the linter to complain about a stutter, i.e.
customerio.CustomerIOError rather then customerio.Error. I've just left this as is.